[Fix]: the 0x02-functions_nested_loops directory files

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int c;
-	int ch = 1, k = 2, r = 0;
+	float ch = 1, k = 2, r = 0;
 
 	printf("1, 2");
 	for (c = 1; c <= 96; c++)
